Remove duplication of decompressed buffer
After a block is decompressed and put into buffer, it's memcpy'd into a new char[] allocated on the heap so it can be referenced by the result Slice. This causes a temporary spike in memory usage until the scope ends (In my testing, sometimes hundreds of Mb!). By allocating the buffer string on the heap from the start, the result Slice can reference buffer's data and take control of its memory management.